Use auth routes as default for upload, merge upload tests
